rename AuthTokenProvider -> FirebaseAuthTokenProvider etc.
