+++ packages/database/src/core/Repo.ts
@@ -26,9 +26,9 @@ import { SparseSnapshotTree } from './SparseSnapshotTree';
 import { SyncTree } from './SyncTree';
 import { SnapshotHolder } from './SnapshotHolder';
 import { stringify } from '@firebase/util';
-import { beingCrawled, each, exceptionGuard, warn, log } from './util/util';
+import { beingCrawled, each, exceptionGuard, warn, log, FIREBASE_DATABASE_EMULATOR_HOST_VAR } from './util/util';
 import { map, isEmpty } from '@firebase/util';
-import { AuthTokenProvider, TokenProvider } from './AuthTokenProvider';
+import { FirebaseAuthTokenProvider, AuthTokenProvider } from './AuthTokenProvider';
 import { EmulatorAuthTokenProvider } from './EmulatorAuthTokenProvider';
 import { StatsManager } from './stats/StatsManager';
 import { StatsReporter } from './stats/StatsReporter';
@@ -38,7 +38,6 @@ import { PersistentConnection } from './PersistentConnection';
 import { ReadonlyRestClient } from './ReadonlyRestClient';
 import { FirebaseApp } from '@firebase/app-types';
 import { RepoInfo } from './RepoInfo';
-import { FIREBASE_DATABASE_EMULATOR_HOST_VAR } from './Constants';
 import { Database } from '../api/Database';
 import { ServerActions } from './ServerActions';
 import { Query } from '../api/Query';
@@ -83,14 +82,14 @@ export class Repo {
     forceRestClient: boolean,
     public app: FirebaseApp
   ) {
-    let authTokenProvider: TokenProvider;
+    let authTokenProvider: AuthTokenProvider;
     if (
       typeof process !== 'undefined' &&
       process.env[FIREBASE_DATABASE_EMULATOR_HOST_VAR]
     ) {
       authTokenProvider = new EmulatorAuthTokenProvider(app);
     } else {
-      authTokenProvider = new AuthTokenProvider(app);
+      authTokenProvider = new FirebaseAuthTokenProvider(app);
     }
 
     this.stats_ = StatsManager.getCollection(repoInfo_);
