+++ packages/database/src/core/RepoManager.ts
@@ -17,9 +17,8 @@
 
 import { FirebaseApp } from '@firebase/app-types';
 import { safeGet } from '@firebase/util';
-import { FIREBASE_DATABASE_EMULATOR_HOST_VAR } from './Constants';
 import { Repo } from './Repo';
-import { fatal } from './util/util';
+import { fatal, FIREBASE_DATABASE_EMULATOR_HOST_VAR } from './util/util';
 import { parseRepoInfo, parseURL } from './util/libs/parser';
 import { validateUrl } from './util/validation';
 import './Repo_transaction';
