+++ packages/database/src/core/util/util.ts
@@ -32,6 +32,17 @@ import { Logger, LogLevel } from '@firebase/logger';
 const logClient = new Logger('@firebase/database');
 
 /**
+ * Environment variable for enabling interaction with the Firebase Realtime Database
+ * emulator. If set, the module will present the endpoint with a fake "owner" credential
+ * (see EmulatorAuthTokenProvider) instead of one belonging to a real account.
+ *
+ * The expected format for this variable is '<HOST>:<PORT>'. The transfer protocol must be
+ * omitted and will default to 'http'.
+ */
+export const FIREBASE_DATABASE_EMULATOR_HOST_VAR =
+  'FIREBASE_DATABASE_EMULATOR_HOST';
+
+/**
  * Returns a locally-unique ID (generated by just incrementing up from 0 each time its called).
  * @type {function(): number} Generated ID.
  */
